refactor(adaptor-naming): Naming the folder “adaptor” is idiotic. It serves no purpose at all except hinting that we should use ports.py, and the subfolders already handle that role perfectly well
