+++ fileorg/file_ops/ports.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
+from pathlib import Path
 from typing import List, Optional
 
 
@@ -93,5 +94,47 @@ class IFileScanner(ABC):
 
         Raises:
             IOError: If the report file cannot be written.
+        """
+        pass
+
+
+@dataclass
+class ParserOutput:
+    """
+    Represents the result of a file parsing operation.
+
+    This data class encapsulates whether parsing succeeded, the extracted text content,
+    and any error or truncation status encountered during the process.
+
+    Attributes:
+        success (bool): Whether the parsing succeeded.
+        content (str): The extracted text content. Empty if parsing failed.
+        is_truncated (bool): Indicates whether the content was truncated (e.g., due to size limits).
+        error (str): Error message if parsing failed, otherwise an empty string.
+    """
+
+    success: bool
+    content: str
+    istruncated: bool
+    error: str
+
+
+class IParser(ABC):
+    """Abstract base for all file parsers.
+
+    Defines the interface that all parsers must follow.
+    Handles common tasks like content truncation and encoding detection.
+    """
+
+    @abstractmethod
+    def parse(self, file_path: Path, char_limit: int) -> ParserOutput:
+        """Extract content from a file.
+
+        Args:
+            file_path: Path to the file to parse
+            char_limit: Maximum characters to extract (prevents memory issues)
+
+        Returns:
+            ParseResult with extracted content or error info
         """
         pass
